Code cleanup from backend to front end

+++ src/main/java/com/example/repository/BrandRepository.java
@@ -10,7 +10,7 @@ import com.example.entity.Brand;
 @Repository
 public interface BrandRepository extends JpaRepository<Brand, Long> {
 
-	List<Brand> findByVehicleTypeIdAndIsDeletedOrderByBrandAsc(long vehicleTypeId, int isDeleted);
+	List<Brand> findByVehicleTypeIdAndIsDeletedOrderByBrandAsc(int vehicleTypeId, int isDeleted);
 	
 	List<Brand> findByBrandIdAndIsDeleted(long brandId, int isDeleted);
